Moved Parser and PathFinder classes under ModuleConfig (task #3495)
Parser and PathFinder classes shouldn't be used by themselves, as they
are only about module configuration anyway.  ModuleConfig is the single
responsibility classes to abstract the details from the consumer of this
functionality.

* Parser files moved under ModuleConfig/Parser
* PathFinder files moved under ModuleConfig/PathFinder
* Parser tests moved under ModuleConfig/Parser
* PathFinder tests moved under ModuleConfig/PathFinder
* Corrected all namespaces and class references
* Adjusted paths to test data in unit tests (one extra level up)
