+++ src/ModuleConfig/Parser/Ini/AbstractIniParser.php
@@ -1,0 +1,8 @@
+<?php
+namespace Qobo\Utils\ModuleConfig\Parser\Ini;
+
+use Qobo\Utils\ModuleConfig\Parser\AbstractParser;
+
+abstract class AbstractIniParser extends AbstractParser
+{
+}
