--- src/Parser/Csv/AbstractCsvParser.php
@@ -1,31 +1,0 @@
-<?php
-namespace Qobo\Utils\Parser\Csv;
-
-use League\Csv\Reader;
-use Qobo\Utils\Parser\AbstractParser;
-
-abstract class AbstractCsvParser extends AbstractParser
-{
-    /**
-     * Mode to use for opening CSV files
-     */
-    protected $open_mode = 'r';
-
-    /**
-     * Get headers from path
-     *
-     * @param string $path Path to file
-     * @return array
-     */
-    public function getHeadersFromPath($path)
-    {
-        $result = [];
-
-        $this->validatePath($path);
-
-        $reader = Reader::createFromPath($path, $this->open_mode);
-        $result = $reader->fetchOne();
-
-        return $result;
-    }
-}
