--- src/Parser/Ini/AbstractIniParser.php
@@ -1,8 +1,0 @@
-<?php
-namespace Qobo\Utils\Parser\Ini;
-
-use Qobo\Utils\Parser\AbstractParser;
-
-abstract class AbstractIniParser extends AbstractParser
-{
-}
