+++ tests/TestCase/ModuleConfig/ModuleConfigTest.php
@@ -4,8 +4,8 @@ namespace Qobo\Utils\Test\TestCase\ModuleConfig;
 use Cake\Core\Configure;
 use PHPUnit_Framework_TestCase;
 use Qobo\Utils\ModuleConfig\ModuleConfig;
-use Qobo\Utils\Parser\Ini\Parser;
-use Qobo\Utils\PathFinder\ConfigPathFinder;
+use Qobo\Utils\ModuleConfig\Parser\Ini\Parser;
+use Qobo\Utils\ModuleConfig\PathFinder\ConfigPathFinder;
 
 class ModuleConfigTest extends PHPUnit_Framework_TestCase
 {
