--- tests/TestCase/Parser/Csv/ViewParserTest.php
@@ -1,18 +1,0 @@
-<?php
-namespace Qobo\Utils\Test\TestCase\Parser\Csv;
-
-use PHPUnit_Framework_TestCase;
-use Qobo\Utils\Parser\Csv\ViewParser;
-
-class ViewParserTest extends PHPUnit_Framework_TestCase
-{
-    public function testParseFromPath()
-    {
-        $file = dirname(dirname(dirname(__DIR__))) . DS . 'data' . DS . 'Modules' . DS . 'Foo' . DS . 'views' . DS . 'view.csv';
-        $parser = new ViewParser();
-        $result = $parser->parseFromPath($file);
-
-        $this->assertTrue(is_array($result), "Parser returned a non-array");
-        $this->assertFalse(empty($result), "Parser returned empty result");
-    }
-}
